% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_tuning.R
\name{adapt_tuning}
\alias{adapt_tuning}
\title{adaptative tuning}
\usage{
adapt_tuning(repli, I, N_geo, within_iter, theta0, sigma, SI, mu0)
}
\arguments{
\item{repli}{number of time the variance of the proposal is tuned (10 tends to be ok)}

\item{within_iter}{iterations for evaluate the accpetance with new proposal variances}

\item{sigma}{original variance to start with}

\item{others}{same as in MCMC_iter function}
}
\description{
tune proposal and give good initial values to start 'proper' MCMC
 try to tune variance toward 20% acceptance
}
\details{
res list of 2 vectors: theta0: posterior samples at the last iterations
                      sigma: new variance for the proposal distribution
}
