% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_incidence.R
\name{ag_incidence}
\alias{ag_incidence}
\title{ag_incidence}
\usage{
ag_incidence(I, delta)
}
\arguments{
\item{I}{matrix of incidence per unit time (nrow: locations, ncol: nb of time points).}

\item{delta}{integer, period over which to aggregate, e.g. if I is daily incidence, and delta is 7, the function return weekly incidence.}
}
\description{
agregate an incidence for periods of delta days, it cuts the most recent days if incidence has not the exact number of days require.
}
