% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Like1.R
\name{Like1}
\alias{Like1}
\title{get likelihood}
\usage{
Like1(lambda, I, R0)
}
\arguments{
\item{I}{matrix of observed incidence, same dimension as lambda}

\item{R0}{vector of reproduction numbers per locations}

\item{lambda:}{'force of infection' matrix (incidence weighted by serial interval),
column number of days, row: number of locations}
}
\description{
get likelihood  
internal to the MCMC
}
\details{
L log likelihood
}
