% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikeNb.R
\name{LikeNb}
\alias{LikeNb}
\title{get likelihood}
\usage{
LikeNb(lambda, I, R0, over_disp)
}
\arguments{
\item{I}{matrix of observed incidence, same dimension as lambda}

\item{R0}{vector of reproduction numbers per locations}

\item{lambda:}{'force of infection' matrix (incidence weighted by serial interval),
column number of days, row: number of locations}
}
\description{
get NegBin likelihood  
internal to the MCMC
}
\details{
L log likelihood
}
