% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_full.R
\name{MCMC_full}
\alias{MCMC_full}
\title{MCMC iterate}
\usage{
MCMC_full(I, N_geo, iter, theta0, s, SI, mu0, repli_adapt, within_iter,
  over_disp = NA)
}
\arguments{
\item{I}{the incidence for the time window during which we assume Rt to be constant.
I is a dataframe, first column are dates then incidence for all locations
nb of row is the size of time widows, dates must be sequential}

\item{N_geo}{integer of  numbers of locations}

\item{iter}{integer, the number of iteration for the MCMC}

\item{theta0}{vector of inital parameters}

\item{s}{variance of proposal distributions (log-normal)}

\item{SI}{Serial interval distribution (see SI_gamma_dist_EpiEstim)}

\item{within_iter}{iterations for evaluate the accpetance with new proposal variances}

\item{mu0:}{initial conidtions to guaranty that if R=1, then we predict the number of cases in the future will stablise at the mean number of cases observed in the time window
mu0 is also used as the mean of the (exponential) prior for intial conditions estimated}

\item{repli}{number of time the variance of the proposal is tuned (10 tends to be ok)}
}
\description{
run the MCMC to sample posterior of R and initial coniditions at each location 
FYI: this is called internally by adapt_tuning
}
\details{
res a list containing 2 matrices: theta: matrix of posterior samples
                     and logL: matrix of associated log-likelihood
}
