% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proj_NegBin_incl_trusted.R
\name{Proj_NegBin_incl_trusted}
\alias{Proj_NegBin_incl_trusted}
\title{Project forward}
\usage{
Proj_NegBin_incl_trusted(Results, Overdispersion, NR_samples, Nsim_per_samples,
  week_forward, N_geo, SI, I)
}
\arguments{
\item{Results}{list conatining the results of the MCMC (e.g. from MCMC_RtProj function), contains
'theta' a matrix for the posterior samples of Rt and initial conditions (ncol: twice number of locations)
with Rts in first column followed by initial conditions.}

\item{Overdispersion}{single real number for Overdispersion}

\item{NR_samples}{integer, number of Rt sampled - has to be smaler than number of posterior samples}

\item{Nsim_per_samples}{integer, number of of simulation per Rt sampled (NR_samples). the total number of 
simulated trajectory is (NR_samples x Nsim_per_samples)}

\item{week_forward}{integer, number of week simulated forward - this include the time window used for inference, so
if 2 weeks were used to infer Rt, and we want projection 2 weeks after that, week_forward=4}

\item{N_geo}{integer, number of locations}

\item{SI}{serial interval, output from SI_gamma_dist_EpiEstim function}

\item{I}{incidence without dates (nrow=nb of days)}
}
\description{
Project forward according to NegBin
}
\details{
I_predict an array of size [Nsim,N_geo,7*weekforward] of simulated incidences
}
