% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_gamma_dist_EpiEstim.R
\name{SI_gamma_dist_EpiEstim}
\alias{SI_gamma_dist_EpiEstim}
\title{SerialInterval}
\usage{
SI_gamma_dist_EpiEstim(mu, cv, SItrunc)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
